Updates to end Sept 2022

+++ loadimage.py
@@ -26,9 +26,9 @@ def loadimage(imfile, flatfile, biasfile):
     except OSError as e:
         raise LoadImErr("Could not open file", e.strerror)
 
-    fd = fd.astype(np.float32)
-    bd = bd.astype(np.float32)
-    imd = imd.astype(np.float32)
+    fd = fd.astype(np.float64)
+    bd = bd.astype(np.float64)
+    imd = imd.astype(np.float64)
 
     fd = trimarrays.trimzeros(trimarrays.trimnan(fd))
     imd, bd = trimarrays.trimto(fd, imd, bd)
@@ -54,9 +54,9 @@ def loadimagehdr(imfile, flatfile, biasfile):
     except OSError as e:
         raise LoadImErr("Could not open file", e.filename, e.strerror)
 
-    fd = fd.astype(np.float32)
-    bd = bd.astype(np.float32)
-    imd = imd.astype(np.float32)
+    fd = fd.astype(np.float64)
+    bd = bd.astype(np.float64)
+    imd = imd.astype(np.float64)
 
     fd = trimarrays.trimzeros(trimarrays.trimnan(fd))
     imd, bd = trimarrays.trimto(fd, imd, bd)
